+++ MoltenVK/MoltenVK/GPUObjects/MVKQueryPool.mm
@@ -316,7 +316,9 @@ MVKOcclusionQueryPool::MVKOcclusionQueryPool(MVKDevice* device,
         VkDeviceSize newBuffLen = min(reqBuffLen, maxBuffLen);
 
         if (reqBuffLen > maxBuffLen) {
-            reportError(VK_ERROR_OUT_OF_DEVICE_MEMORY, "vkCreateQueryPool(): Each query pool can support a maximum of %d queries.", uint32_t(newBuffLen / kMVKQuerySlotSizeInBytes));
+			reportError(VK_ERROR_OUT_OF_DEVICE_MEMORY,
+						"vkCreateQueryPool(): Each occlusion query pool can support a maximum of %d queries.",
+						uint32_t(newBuffLen / kMVKQuerySlotSizeInBytes));
         }
 
         NSUInteger mtlBuffLen = mvkAlignByteCount(newBuffLen, _device->_pMetalFeatures->mtlBufferAlignment);
@@ -356,9 +358,9 @@ void MVKGPUCounterQueryPool::initMTLCounterSampleBuffer(const VkQueryPoolCreateI
 		NSError* err = nil;
 		_mtlCounterBuffer = [getMTLDevice() newCounterSampleBufferWithDescriptor: tsDesc error: &err];
 		if (err) {
-			setConfigurationResult(reportError(VK_ERROR_INITIALIZATION_FAILED,
-											   "Could not create MTLCounterSampleBuffer for query pool of type %s. Reverting to emulated behavior. (Error code %li): %s",
-											   queryTypeName, (long)err.code, err.localizedDescription.UTF8String));
+			reportError(VK_ERROR_OUT_OF_DEVICE_MEMORY,
+						"Could not create MTLCounterSampleBuffer of size %llu, for %d queries, in query pool of type %s. Reverting to emulated behavior. (Error code %li): %s",
+						(VkDeviceSize)pCreateInfo->queryCount * kMVKQuerySlotSizeInBytes, pCreateInfo->queryCount, queryTypeName, (long)err.code, err.localizedDescription.UTF8String);
 		}
 	}
 };